minor changes to find the gradient bug. Not succesful.

+++ Assignment 1/Initialize.m
@@ -1,6 +1,6 @@
 function [W, b] = Initialize(K, d, initType)
-  W = randn(K,d);
-  b = randn(K,1);
+  W = double(randn(K,d));
+  b = double(randn(K,1));
   if nargin < 3
     W = W*sqrt(0.1);
     b = b*sqrt(0.1);
